requirements.txt
ctransformers
sentence-transformers
pinecone-client
langchain
langchain-community
langchain_pinecone
langchain-huggingface
fastapi
uvicorn
streamlit
pypdf
python-dotenv
Werkzeug
-e .